Jump on Mobile fix/Platform Spawning Issues
Jumping now works when pressing the button on mobile (did not
previously, but did on a PC).

Also working on Platform Spawning issues so they are not spawned on top
of each other, too many at once, or too close eventually. Some of these
fixes will most likely come further in development.