Updates 9/21 MainMenu/Landscape Mode
Added a basic main menu with a button to play the game.

Modified the "GameOverScene" to be a little more clean. Still need to
grab the score (or distance traveled right now) to display after the
game is over.

Added a line of code to the start of each scene (in any script usually
attached to the camera in that scene) so it stays locked in Landscape
mode to the left (turn phone left when holding it normally).

- DR